Add Elixir 1.6 to travis matrix
I also trimmed down the number of builds so that we only run a single
build for each version of Elixir.

I found 2 credo issues and fixed them. I would be in favor of running
credo and dialyzer in travis, but would like a second opinion on that.

Also updated some dependencies